+++ swap/src/asb/rpc/server.rs
@@ -345,6 +345,24 @@ impl AsbApiServer for RpcImpl {
         Ok(())
     }
 
+    async fn set_external_bitcoin_redeem_address(
+        &self,
+        address: Option<String>,
+    ) -> Result<(), ErrorObjectOwned> {
+        let network = self.bitcoin_wallet.network();
+        let address = address
+            .map(|s| bitcoin_wallet::bitcoin_address::parse_and_validate_network(&s, network))
+            .transpose()
+            .into_json_rpc_result()?;
+
+        self.event_loop_service
+            .set_external_bitcoin_redeem_address(address)
+            .await
+            .into_json_rpc_result()?;
+
+        Ok(())
+    }
+
     async fn get_current_quote(&self) -> Result<QuoteResponse, ErrorObjectOwned> {
         let quote = self
             .event_loop_service